submit iTunesConnect for Test

+++ YF002/YF002/library/DropView/LMComBox/LMComBoxView.m
@@ -198,7 +198,7 @@
     [self tapAction];
     if([_delegate respondsToSelector:@selector(selectAtIndex:inCombox:)])
     {
-        [_delegate selectAtIndex:indexPath.row inCombox:self];
+        [_delegate selectAtIndex:(NSInteger)indexPath.row inCombox:self];
     }
     [self performSelector:@selector(deSelectedRow) withObject:nil afterDelay:0.2];
 }
